+++ modules/cms/src/frontend/blocks/DevBlock.php
@@ -40,7 +40,7 @@ class DevBlock extends TwigBlock
         return '';
     }
     
-    public function getBlockGroup()
+    public function blockGroup()
     {
         return DevelopmentGroup::className();
     }