+++ modules/cms/src/frontend/blocks/ImageTextBlock.php
@@ -29,7 +29,7 @@ class ImageTextBlock extends TwigBlock
         return Module::t('block_image_text_name');
     }
     
-    public function getBlockGroup()
+    public function blockGroup()
     {
         return MediaGroup::class;
     }