+++ modules/cms/tests/src/base/BlockTest.php
@@ -56,9 +56,8 @@ class BlockTest extends CmsFrontendTestCase
 
         $this->assertEquals('TestBlock.twig', $block->getViewFileName('twig'));
         $this->assertEquals('twig-frontend', $block->renderFrontend());
-        $this->assertEquals('<i class="material-icons">test-icon</i> <span>Test</span>', $block->getFullName());
 
-        foreach ($block->getVars() as $var) {
+        foreach ($block->getVarsExport() as $var) {
             $this->assertArrayHasKey('id', $var);
             $this->assertArrayHasKey('var', $var);
             $this->assertArrayHasKey('label', $var);
@@ -68,7 +67,7 @@ class BlockTest extends CmsFrontendTestCase
             $this->assertArrayHasKey('initvalue', $var);
         }
 
-        foreach ($block->getCfgs() as $var) {
+        foreach ($block->getCfgsExport() as $var) {
             $this->assertArrayHasKey('id', $var);
             $this->assertArrayHasKey('var', $var);
             $this->assertArrayHasKey('label', $var);
@@ -100,7 +99,7 @@ class BlockTest extends CmsFrontendTestCase
     {
         $block = new FailureBlock();
         // will throw Exception:  Required attributes in config var element is missing. var, label and type are required.
-        $block->getVars();
+        $block->getVarsExport();
     }
 
     public function testGetterSetter()