+++ modules/cms/tests/src/injectors/ActiveQueryCheckboxInjectorTest.php
@@ -28,7 +28,7 @@ class ActiveQueryCheckboxInjectorTest extends CmsFrontendTestCase
         $injector = new ActiveQueryCheckboxInjector(['query' => NavItem::find(), 'varName' => 'test', 'varLabel' => 'test label', 'context' => $block]);
         $injector->setup();
         
-        $vars = $block->getVars();
+        $vars = $block->getVarsExport();
         
         $v = $vars[0];
         
@@ -50,7 +50,7 @@ class ActiveQueryCheckboxInjectorTest extends CmsFrontendTestCase
         $injector = new ActiveQueryCheckboxInjector(['query' => NavItem::find()->select(['title']), 'varName' => 'test', 'varLabel' => 'test label', 'context' => $block]);
         $injector->setup();
     
-        $vars = $block->getVars();
+        $vars = $block->getVarsExport();
     
         $v = $vars[0];